+++ packages/forklift-console-plugin/locales/en/plugin__forklift-console-plugin.json
@@ -361,6 +361,7 @@
   "Plan name": "Plan name",
   "Plans": "Plans",
   "Plans for virtualization": "Plans for virtualization",
+  "Plans wizard": "Plans wizard",
   "Please choose a NetworkAttachmentDefinition for data transfer.": "Please choose a NetworkAttachmentDefinition for data transfer.",
   "Please choose a NetworkAttachmentDefinition for default data transfer.": "Please choose a NetworkAttachmentDefinition for default data transfer.",
   "Please choose a target namespace for the migrated virtual machines.": "Please choose a target namespace for the migrated virtual machines.",
@@ -495,6 +496,7 @@
   "The current certificate does not match the certificate fetched from URL. Manually validate the fingerprint before proceeding.": "The current certificate does not match the certificate fetched from URL. Manually validate the fingerprint before proceeding.",
   "The edit mappings button is disabled when the plan status does not enable editing.": "The edit mappings button is disabled when the plan status does not enable editing.",
   "The edit virtual machines button is disabled when the plan status does not enable editing.": "The edit virtual machines button is disabled when the plan status does not enable editing.",
+  "The following VMs do not exist on the source provider and will be removed from the plan": "The following VMs do not exist on the source provider and will be removed from the plan",
   "The interval in minutes for precopy. Default value is 60.": "The interval in minutes for precopy. Default value is 60.",
   "The interval in seconds for snapshot pooling. Default value is 10.": "The interval in seconds for snapshot pooling. Default value is 10.",
   "The limit for CPU usage by the controller, specified in milliCPU. Default value is 500m.": "The limit for CPU usage by the controller, specified in milliCPU. Default value is 500m.",
