+++ packages/forklift-console-plugin/src/modules/Plans/utils/index.ts
@@ -1,5 +1,7 @@
 // @index(['./*', /style/g], f => `export * from '${f.path}';`)
 export * from './constants';
 export * from './helpers';
+export * from './patchPlanMappingsData';
+export * from './patchPlanSpecVms';
 export * from './types';
 // @endindex
